refactor: rename db to scriptInfoDb for more clear
